🔵refactor: AnswerCommand를 domain에서 controller로 이동
