+++ src/main/java/controller/AnswerCommand.java
@@ -1,4 +1,4 @@
-package domain;
+package controller;
 
 import java.util.Arrays;
 import java.util.Objects;
